+++ lace/src/lib.rs
@@ -190,10 +190,9 @@ pub use config::EngineUpdateConfig;
 pub use interface::{
     update_handler, utils, AppendStrategy, BuildEngineError,
     ConditionalEntropyType, DatalessOracle, Engine, EngineBuilder, Given,
-    HasData, HasStates, ImputeUncertaintyType, InsertDataActions, InsertMode,
-    Metadata, MiComponents, MiType, Oracle, OracleT, OverwriteMode,
-    PredictUncertaintyType, Row, RowSimilarityVariant, SupportExtension, Value,
-    WriteMode,
+    HasData, HasStates, InsertDataActions, InsertMode, Metadata, MiComponents,
+    MiType, Oracle, OracleT, OverwriteMode, Row, RowSimilarityVariant,
+    SupportExtension, Value, WriteMode,
 };
 
 pub mod error {