+++ pylace/lace/analysis.py
@@ -1,13 +1,20 @@
-"""Tools for analysis of CrossCat results in Lace."""
+"""Tools for analysis of probabilistic cross-categorization results in Lace."""
 
 
 import enum
 import itertools as it
 from copy import deepcopy
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any, Optional, Tuple, Union
 
 import polars as pl
 from tqdm import tqdm
+
+from lace import utils
+
+ABLATIVE_ERR = "ablative-err"
+ABLATIVE_DIST = "ablative-dist"
+PRED_EXPLAIN_METHODS = [ABLATIVE_ERR, ABLATIVE_DIST]
+
 
 if TYPE_CHECKING:
     from lace import Engine
@@ -34,6 +41,7 @@ class HoldOutFunc(enum.Enum):
 
     NegLogp = 0
     Inconsistency = 1
+    Uncertainty = 2
 
     def __repr__(self) -> str:
         """Return the canonical string representation of the object."""
@@ -41,6 +49,8 @@ class HoldOutFunc(enum.Enum):
             return "-logp"
         elif self == HoldOutFunc.Inconsistency:
             return "inconsistency"
+        elif self == HoldOutFunc.Uncertainty:
+            return "uncertainty"
         else:
             raise NotImplementedError
 
@@ -59,6 +69,8 @@ def _held_out_compute(
             return None
     elif fn == HoldOutFunc.Inconsistency:
         return engine.inconsistency(values, given=given)
+    elif fn == HoldOutFunc.Uncertainty:
+        return engine.predict(values, given=given)[1]
     else:
         raise ValueError(f"Invalid computation `{fn}`")
 
@@ -423,6 +435,565 @@ def held_out_inconsistency(
         quiet=quiet,
     )
     return res
+
+
+def held_out_uncertainty(
+    engine: "Engine",
+    target: Union[str, int],
+    given: dict[Union[str, int], Any],
+    quiet: bool = False,
+    greedy: bool = True,
+) -> pl.DataFrame:
+    r"""
+    Compute prediction uncertainty while sequentially dropping given conditions.
+
+    Parameters
+    ----------
+    engine: Engine
+        The Engine used to compute inconsistency
+    target: str or int
+        The target column for prediction
+    given: dict[index, value], optional
+        A dictionary mapping column indices/name to values, which specifies
+        conditions on the observations.
+    quiet: bool
+        Prevent the display of a progress bar.
+    greedy: bool
+        Use a greedy algorithm which is faster but may be less optimal.
+
+    Returns
+    -------
+    polars.DataFrame
+        A DataFrame with a 'feature' column and a uncertainty column.
+
+    Examples
+    --------
+    >>> import polars as pl
+    >>> from lace.examples import Satellites
+    >>> from lace.analysis import held_out_uncertainty
+    >>> satellites = Satellites()
+    >>> given = (
+    ...     satellites.df.to_pandas()
+    ...     .set_index("ID")
+    ...     .loc["Intelsat 903", :]
+    ...     .dropna()
+    ...     .to_dict()
+    ... )
+    >>> period = given.pop("Period_minutes")
+    >>> held_out_uncertainty(
+    ...     satellites,
+    ...     "Period_minutes",
+    ...     given,
+    ...     quiet=True,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    shape: (19, 3)
+    ┌──────────────────────────────────┬─────────────────────────┬───────────┐
+    │ feature_rmed                     ┆ HoldOutFunc.Uncertainty ┆ keys_rmed │
+    │ ---                              ┆ ---                     ┆ ---       │
+    │ list[str]                        ┆ f64                     ┆ i64       │
+    ╞══════════════════════════════════╪═════════════════════════╪═══════════╡
+    │ null                             ┆ 0.43212                 ┆ 0         │
+    │ ["Perigee_km"]                   ┆ 0.43212                 ┆ 1         │
+    │ ["Class_of_Orbit"]               ┆ 0.43212                 ┆ 2         │
+    │ ["Source_Used_for_Orbital_Data"] ┆ 0.431921                ┆ 3         │
+    │ …                                ┆ …                       ┆ …         │
+    │ ["Country_of_Operator"]          ┆ 0.054156                ┆ 15        │
+    │ ["Country_of_Contractor"]        ┆ 0.06069                 ┆ 16        │
+    │ ["Dry_Mass_kg"]                  ┆ 0.139502                ┆ 17        │
+    │ ["Inclination_radians"]          ┆ 0.089026                ┆ 18        │
+    └──────────────────────────────────┴─────────────────────────┴───────────┘
+
+    If we don't want to use the greedy search, we can enumerate, but we need to
+    be mindful that the number of conditions we must enumerate over is 2^n
+
+    >>> keys = sorted(list(given.keys()))
+    >>> _ = [given.pop(c) for c in keys[-10:]]
+    >>> held_out_uncertainty(
+    ...     satellites,
+    ...     "Period_minutes",
+    ...     given,
+    ...     quiet=True,
+    ...     greedy=False,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    shape: (9, 3)
+    ┌───────────────────────────────────┬─────────────────────────┬───────────┐
+    │ feature_rmed                      ┆ HoldOutFunc.Uncertainty ┆ keys_rmed │
+    │ ---                               ┆ ---                     ┆ ---       │
+    │ list[str]                         ┆ f64                     ┆ i64       │
+    ╞═══════════════════════════════════╪═════════════════════════╪═══════════╡
+    │ null                              ┆ 0.445501                ┆ 0         │
+    │ ["Expected_Lifetime"]             ┆ 0.437647                ┆ 1         │
+    │ ["Apogee_km", "Eccentricity"]     ┆ 0.05561                 ┆ 2         │
+    │ ["Apogee_km", "Country_of_Operat… ┆ 0.055283                ┆ 3         │
+    │ …                                 ┆ …                       ┆ …         │
+    │ ["Apogee_km", "Country_of_Operat… ┆ 0.057624                ┆ 5         │
+    │ ["Apogee_km", "Country_of_Contra… ┆ 0.0595                  ┆ 6         │
+    │ ["Apogee_km", "Country_of_Contra… ┆ 0.077359                ┆ 7         │
+    │ ["Apogee_km", "Class_of_Orbit", … ┆ 0.089026                ┆ 8         │
+    └───────────────────────────────────┴─────────────────────────┴───────────┘
+    """
+    search = (
+        HoldOutSearchMethod.Greedy if greedy else HoldOutSearchMethod.Enumerate
+    )
+
+    res = _held_out_base(
+        engine,
+        HoldOutFunc.Uncertainty,
+        search,
+        target,
+        deepcopy(given),
+        quiet=quiet,
+    )
+    return res
+
+
+def _attributable_holdout(
+    engine: "Engine",
+    fn: HoldOutFunc,
+    values,
+    given: dict[Union[str, int], Any],
+    quiet: bool = False,
+    greedy: bool = True,
+):
+    search = (
+        HoldOutSearchMethod.Greedy if greedy else HoldOutSearchMethod.Enumerate
+    )
+
+    fn_str = str(fn)
+
+    res = _held_out_base(
+        engine, fn, search, values, deepcopy(given), quiet=quiet
+    )
+
+    n_holdouts = res.shape[0]
+
+    if n_holdouts < 2:
+        return 0.0, res
+
+    area = 0.0
+    fn_max = res[fn_str][0]
+
+    fn_a = res[fn_str][1]
+    area += (fn_a / fn_max) / 2.0
+
+    n = 2
+
+    for i in range(2, n_holdouts):
+        fn_b = res[fn_str][i]
+
+        if fn_b > fn_max:
+            break
+
+        area += ((fn_max - fn_b) / fn_max) / 2.0
+        fn_a = fn_b
+        n += 1
+
+    if n <= 2:
+        return 0, res
+
+    return area / n, res
+
+
+def attributable_inconsistency(
+    engine: "Engine",
+    values,
+    given: dict[Union[str, int], Any],
+    quiet: bool = False,
+    greedy: bool = True,
+) -> Tuple[float, pl.DataFrame]:
+    r"""
+    Determine what fraction of inconsistency is attributable.
+
+    The fraction will be higher if dropping fewer predictor reduces
+    inconsistency quickly. The fraction will be 1 if one predictor drops
+    inconsistency to zero (this is unlikely to ever occur). The fraction will
+    be 0 if dropping predictors has no effect.
+
+    Parameters
+    ----------
+    engine: Engine
+        The Engine used to compute inconsistency
+    values: polars or pandas DataFrame or Series
+        The values over which to compute the inconsistency. Each row of the
+        DataFrame, or each entry of the Series, is an observation. Column
+        names (or the Series name) should correspond to names of features in
+        the table.
+    given: dict[index, value], optional
+        A dictionary mapping column indices/name to values, which specifies
+        conditions on the observations.
+    quiet: bool
+        Prevent the display of a progress bar.
+    greedy: bool
+        Use a greedy algorithm which is faster but may be less optimal.
+
+    Returns
+    -------
+    float
+        The fraction [0, 1] of the inconsistency that is attributable
+    polars.DataFrame
+        The result of held_out_inconsistency
+
+    Examples
+    --------
+    >>> import polars as pl
+    >>> from lace.examples import Satellites
+    >>> from lace.analysis import attributable_inconsistency
+    >>> satellites = Satellites()
+    >>> given = (
+    ...     satellites.df.to_pandas()
+    ...     .set_index("ID")
+    ...     .loc["Intelsat 903", :]
+    ...     .dropna()
+    ...     .to_dict()
+    ... )
+    >>> period = given.pop("Period_minutes")
+    >>> frac, df = attributable_inconsistency(
+    ...     satellites,
+    ...     pl.Series("Period_minutes", [period]),
+    ...     given,
+    ...     quiet=True,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    >>> frac
+    0.2930260843667006
+    """
+
+    return _attributable_holdout(
+        engine,
+        HoldOutFunc.Inconsistency,
+        values,
+        given,
+        quiet=quiet,
+        greedy=greedy,
+    )
+
+
+def attributable_neglogp(
+    engine: "Engine",
+    values,
+    given: dict[Union[str, int], Any],
+    quiet: bool = False,
+    greedy: bool = True,
+) -> Tuple[float, pl.DataFrame]:
+    r"""
+    Determine what fraction of surprisal (-log p) is attributable.
+
+    The fraction will be higher if dropping fewer predictor reduces surprisal
+    quickly. The fraction will be 1 if one predictor drops surprisal to zero
+    (this can never occur). The fraction will be 0 if dropping predictors has
+    no effect.
+
+    Parameters
+    ----------
+    engine: Engine
+        The Engine used to compute inconsistency
+    values: polars or pandas DataFrame or Series
+        The values over which to compute the -log p. Each row of the
+        DataFrame, or each entry of the Series, is an observation. Column
+        names (or the Series name) should correspond to names of features in
+        the table.
+    given: dict[index, value], optional
+        A dictionary mapping column indices/name to values, which specifies
+        conditions on the observations.
+    quiet: bool
+        Prevent the display of a progress bar.
+    greedy: bool
+        Use a greedy algorithm which is faster but may be less optimal.
+
+    Returns
+    -------
+    float
+        The fraction [0, 1] of the surprisal that is attributable
+    polars.DataFrame
+        The result of held_out_neglogp
+
+    Examples
+    --------
+    >>> import polars as pl
+    >>> from lace.examples import Satellites
+    >>> from lace.analysis import attributable_neglogp
+    >>> satellites = Satellites()
+    >>> given = (
+    ...     satellites.df.to_pandas()
+    ...     .set_index("ID")
+    ...     .loc["Intelsat 903", :]
+    ...     .dropna()
+    ...     .to_dict()
+    ... )
+    >>> period = given.pop("Period_minutes")
+    >>> frac, df =  attributable_neglogp(
+    ...     satellites,
+    ...     pl.Series("Period_minutes", [period]),
+    ...     given,
+    ...     quiet=True,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    >>> frac
+    0.29302608436670047
+    """
+
+    return _attributable_holdout(
+        engine,
+        HoldOutFunc.NegLogp,
+        values,
+        given,
+        quiet=quiet,
+        greedy=greedy,
+    )
+
+
+def attributable_uncertainty(
+    engine: "Engine",
+    target: Union[str, int],
+    given: dict[Union[str, int], Any],
+    quiet: bool = False,
+    greedy: bool = True,
+) -> Tuple[float, pl.DataFrame]:
+    r"""
+    Determine what fraction of uncertainty is attributable.
+
+    The fraction will be higher if dropping fewer predictor reduces uncertainty
+    quickly. The fraction will be 1 if one predictor drops uncertainty to zero
+    (this is unlikely). The fraction will be 0 if dropping predictors has no
+    effect.
+
+    Parameters
+    ----------
+    engine: Engine
+        The Engine used to compute inconsistency
+    target: str or int
+        The prediction target
+    given: dict[index, value], optional
+        A dictionary mapping column indices/name to values, which specifies
+        conditions on the observations.
+    quiet: bool
+        Prevent the display of a progress bar.
+    greedy: bool
+        Use a greedy algorithm which is faster but may be less optimal.
+
+    Returns
+    -------
+    float
+        The fraction [0, 1] of the uncertainty that is attributable
+    polars.DataFrame
+        The result of held_out_uncertainty
+
+    Examples
+    --------
+    >>> import polars as pl
+    >>> from lace.examples import Satellites
+    >>> from lace.analysis import attributable_uncertainty
+    >>> satellites = Satellites()
+    >>> given = (
+    ...     satellites.df.to_pandas()
+    ...     .set_index("ID")
+    ...     .loc["Intelsat 903", :]
+    ...     .dropna()
+    ...     .to_dict()
+    ... )
+    >>> period = given.pop("Period_minutes")
+    >>> frac, df =  attributable_uncertainty(
+    ...     satellites,
+    ...     "Period_minutes",
+    ...     given,
+    ...     quiet=True,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    >>> frac
+    0.1814171785207335
+    """
+
+    return _attributable_holdout(
+        engine,
+        HoldOutFunc.Uncertainty,
+        target,
+        given,
+        quiet=quiet,
+        greedy=greedy,
+    )
+
+
+def _explain_ablative_err(
+    engine: "Engine",
+    target: Union[int, str],
+    given: dict[Union[str, int], Any],
+):
+    xs = utils.predict_xs(engine, target, None, mass=0.995)
+
+    ftype = engine.ftype(target)
+    if ftype == "Categorial":
+        norm = 1.0 / len(xs)
+    elif ftype in ("Continuous", "Count"):
+        norm = xs[1] - xs[0]
+    else:
+        raise ValueError(f"Unupported FType `{ftype}`")
+
+    baseline = engine.logp(xs, given).exp()
+
+    cols = []
+    imps = []
+    for k in list(given.keys()):
+        # remove the value
+        val = given.pop(k)
+        ps = engine.logp(xs, given).exp()
+
+        err = (ps - baseline).sum() * norm
+
+        # put the value back
+        given[k] = val
+
+        cols.append(k)
+        imps.append(err)
+
+    return cols, imps
+
+
+def _explain_ablative_dist(
+    engine: "Engine",
+    target: Union[int, str],
+    given: dict[Union[str, int], Any],
+):
+    ftype = engine.ftype(target)
+    if ftype != "Continuous":
+        msg = (
+            "`ablative-dist` explanation only valid for Continuous targets"
+            f" but target `{target}` is `{ftype}`"
+        )
+        raise ValueError(msg)
+
+    cols = []
+    imps = []
+    for k in list(given.keys()):
+        baseline = engine.predict(target, given, with_uncertainty=False)
+
+        # remove the value
+        val = given.pop(k)
+
+        pred = engine.predict(target, given, with_uncertainty=False)
+
+        # put the value back
+        given[k] = val
+
+        cols.append(k)
+        imps.append(pred - baseline)
+
+    return cols, imps
+
+
+def explain_prediction(
+    engine: "Engine",
+    target: Union[int, str],
+    given: dict[Union[str, int], Any],
+    *,
+    method: Optional[str] = None,
+):
+    """
+    Explain the relevance of each predictor when predicting a target.
+
+    Parameters
+    ----------
+    engine: lace.Engine
+        The source engine
+    target: str, int
+        The target variable -- the variable to predict
+    given: Dict[index, value], optional
+        A dictionary mapping column indices/name to values, which specifies
+        conditions on the observations.
+    method: str, optional
+        The method to use for explanation:
+        * 'ablative-err' (default): computes the different between p(y|X) and
+          p(x|X - xᵢ) for each predictor xᵢ in the `given`, X.
+        * 'ablative-dist': computed the error between the predictions (argmax)
+          of p(y|X) and p(x|X - xᵢ) for each predictor xᵢ in the `given`, X. Note
+          that this method does not support categorical targets.
+
+    Returns
+    -------
+    cols: List[str]
+        The column names associated with each importance
+    imps: List[float]
+        The list of importances for each column
+
+    Examples
+    --------
+    >>> import polars as pl
+    >>> from lace.examples import Satellites
+    >>> from lace.analysis import explain_prediction
+    >>> engine = Satellites()
+
+    Define a target
+
+    >>> target = 'Period_minutes'
+
+    We'll use a row from the data
+
+    >>> row = engine[5, :].to_dicts()[0]
+    >>> ix = row.pop('index')
+    >>> _ = row.pop(target)
+    >>> given = { k: v for k, v in row.items() if v is not None }
+
+    The default importance method, 'ablative-err', measures the error between
+    the baseline predictive distribution, and the distribution when a predictor
+    is dropped.
+
+    >>> cols, imps = explain_prediction(
+    ...     engine,
+    ...     target,
+    ...     given,
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    >>> pl.DataFrame({'col': cols, 'imp': imps})
+    shape: (18, 2)
+    ┌──────────────────────────────┬─────────────┐
+    │ col                          ┆ imp         │
+    │ ---                          ┆ ---         │
+    │ str                          ┆ f64         │
+    ╞══════════════════════════════╪═════════════╡
+    │ Country_of_Operator          ┆ 3.5216e-16  │
+    │ Users                        ┆ -3.1668e-14 │
+    │ Purpose                      ┆ -9.5636e-14 │
+    │ Class_of_Orbit               ┆ -1.8263e-15 │
+    │ …                            ┆ …           │
+    │ Launch_Site                  ┆ -2.8416e-15 │
+    │ Launch_Vehicle               ┆ 1.0704e-14  │
+    │ Source_Used_for_Orbital_Data ┆ -3.9301e-15 │
+    │ Inclination_radians          ┆ -9.6259e-15 │
+    └──────────────────────────────┴─────────────┘
+
+    Get the importances using the 'ablative-dist' method, which measures how
+    much the prediction would change if a predictor was dropped.
+
+    >>> cols, imps = explain_prediction(
+    ...     engine,
+    ...     target,
+    ...     given,
+    ...     method='ablative-dist'
+    ... )  # doctest: +NORMALIZE_WHITESPACE
+    >>> pl.DataFrame({'col': cols, 'imp': imps})
+    shape: (18, 2)
+    ┌──────────────────────────────┬───────────┐
+    │ col                          ┆ imp       │
+    │ ---                          ┆ ---       │
+    │ str                          ┆ f64       │
+    ╞══════════════════════════════╪═══════════╡
+    │ Country_of_Operator          ┆ -0.000109 │
+    │ Users                        ┆ 0.081289  │
+    │ Purpose                      ┆ 0.18938   │
+    │ Class_of_Orbit               ┆ 0.000133  │
+    │ …                            ┆ …         │
+    │ Launch_Site                  ┆ 0.003411  │
+    │ Launch_Vehicle               ┆ -0.018817 │
+    │ Source_Used_for_Orbital_Data ┆ 0.001454  │
+    │ Inclination_radians          ┆ 0.057333  │
+    └──────────────────────────────┴───────────┘
+    """
+    if method is None:
+        method = ABLATIVE_ERR
+
+    if method == ABLATIVE_ERR:
+        return _explain_ablative_err(engine, target, given)
+    elif method == ABLATIVE_DIST:
+        return _explain_ablative_dist(engine, target, given)
+    else:
+        raise ValueError(
+            f"Invalid method `{method}`, valid methods are {PRED_EXPLAIN_METHODS}"
+        )
 
 
 if __name__ == "__main__":